Abstract
Owing to the dramatic progress in technologies in the field of synthetic
biology, it became easy to synthesize DNAs and proteins in test tubes.
Depending on these developed techniques, we are thinking if it is
possible to create an artificial cell of the simplest compositions,
because cells comprises DNAs and proteins. A cell can be simply divided
into two parts, a membrane structure as the compartment and a chemical
mixture in which the biochemical reactions (transcription and
translation) occur. In this study, we adopt a phospholipid bilayer
structure as the artificial membrane, and a cell-free protein synthesis
system as the biochemical reactions in the membrane (Figure 1). The
artificial membrane is labeled by fluorescent pigment and observed by a
fluorescence microscope. Transcription and translation are evaluated by
activity of synthesized enzymes. The success of protein synthesis in the
membrane structure is considered as the success of the mimicking an
artificial cells in early stage. In the future, we plan to design an
environment-responsive artificial cell by introducing gene switch into
the artificial cell.